feat: message project co-members from /messages (#770) (#948)
Makes the #768 authorization layer and the #769 API reachable. Two mentees on
the same project, with no mentorship between them, can now find each other and
start a DM; the inbox was built purely from mentorshipRelation.findMany, so they
were invisible to one another.

A "New chat" picker on /messages lists project co-members, derived on the server
from ProjectMember so the client never decides who is messageable.
/messages/c/[conversationId] renders conversations; rather than duplicate ~430
lines of message UI, the thread view moved to MessageThreadView and both routes
are thin wrappers over it.

Losing the shared project now makes a DM read-only rather than unreachable.
Reading stays participant-based and permanent so history does not vanish, but
posting is re-checked by canPostToConversation(): POST /api/messages returns 403
and GET returns a canPost flag so the thread shows a read-only notice. Without
this, participation alone would have kept a removed member writing indefinitely.

e2e/project-dm.spec.ts covers the flow end to end. Not tagged @smoke, and it was
not executed locally (no database in the dev container), so the scheduled full
suite is where it first runs.

Closes #770
