+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "internship-crm",
-  "version": "0.26.2-beta",
+  "version": "0.27.0-beta",
   "private": true,
   "license": "AGPL-3.0-or-later",
   "prisma": {
