+++ src/app/messages/[relationId]/page.tsx
@@ -1,428 +1,12 @@
 'use client';
 
-import { use, useCallback, useEffect, useRef, useState } from 'react';
-import { useSession } from 'next-auth/react';
-import { FileText, Download, MoreVertical, Check, CheckCheck, SmilePlus } from 'lucide-react';
-import { Card } from '@/components/ui/Card';
-import { Button } from '@/components/ui/Button';
-import { useT, useLocale } from '@/i18n/client';
-import { formatDateTime } from '@/lib/relativeTime';
-import {
-  MessageComposer,
-  PendingAttachmentList,
-  type PendingMessageAttachment,
-} from '@/components/MessageThread';
+import { use } from 'react';
+import { MessageThreadView } from '@/components/MessageThreadView';
 
-interface Attachment {
-  id: string;
-  filename: string;
-  contentType: string;
-  size: number;
-}
-interface Msg {
-  id: string;
-  senderId: string;
-  body: string;
-  channel: 'IN_APP' | 'EMAIL';
-  readAt: string | null;
-  createdAt: string;
-  editedAt: string | null;
-  deleted: boolean;
-  attachments: Attachment[];
-  reactions: { emoji: string; count: number; mine: boolean }[];
-}
-
-// Fixed reaction set (mirrors the server's REACTION_EMOJIS).
-const REACTIONS = ['👍', '❤️', '😂', '😮', '🎉'] as const;
-interface Party { id: string; fullName: string }
-
+// A mentorship thread, addressed by its relation id. The thread UI itself is
+// shared with the conversation route (/messages/c/[conversationId]) — see
+// MessageThreadView.
 export default function ThreadPage({ params }: { params: Promise<{ relationId: string }> }) {
   const { relationId } = use(params);
-  const t = useT();
-  const locale = useLocale();
-  const { data: session } = useSession();
-  const myId = session?.user?.id;
-  const [messages, setMessages] = useState<Msg[]>([]);
-  const [mentor, setMentor] = useState<Party | null>(null);
-  const [mentee, setMentee] = useState<Party | null>(null);
-  const [body, setBody] = useState('');
-  // Pending attachments (picked files + pasted images), each with an object URL
-  // for an instant thumbnail/preview. Uploaded with the message on send.
-  const [attachments, setAttachments] = useState<PendingMessageAttachment[]>([]);
-  const [attachError, setAttachError] = useState('');
-  const [loading, setLoading] = useState(true);
-  const [sending, setSending] = useState(false);
-  const [forbidden, setForbidden] = useState(false);
-  // Per-message edit/delete state.
-  const [menuId, setMenuId] = useState<string | null>(null);
-  const [reactId, setReactId] = useState<string | null>(null);
-  const [editId, setEditId] = useState<string | null>(null);
-  const [editBody, setEditBody] = useState('');
-  const [rowBusy, setRowBusy] = useState(false);
-  // Per-user (per-device) composer preference: when on, Enter sends and
-  // Shift+Enter inserts a newline; when off (default), Enter is a newline and
-  // Shift+Enter sends. Persisted in localStorage so it sticks for this user.
-  const [enterToSend, setEnterToSend] = useState(false);
-  const endRef = useRef<HTMLDivElement>(null);
-  const fileRef = useRef<HTMLInputElement>(null);
-
-  // Load the saved Enter-to-send preference once on mount.
-  useEffect(() => {
-    setEnterToSend(localStorage.getItem('messages-enter-to-send') === '1');
-  }, []);
-  const toggleEnterToSend = () => {
-    setEnterToSend((prev) => {
-      const next = !prev;
-      try { localStorage.setItem('messages-enter-to-send', next ? '1' : '0'); } catch { /* ignore */ }
-      return next;
-    });
-  };
-
-  // Enter/Shift+Enter behaviour depends on the preference above; ArrowUp on an
-  // empty box edits your last message (WhatsApp/Slack/Telegram style).
-  const onComposerKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-    if (e.nativeEvent.isComposing) return;
-    if (e.key === 'ArrowUp' && !body && editId === null) {
-      const lastMine = [...messages].reverse().find((m) => m.senderId === myId && !m.deleted);
-      if (lastMine) {
-        e.preventDefault();
-        setMenuId(null);
-        setReactId(null);
-        setEditId(lastMine.id);
-        setEditBody(lastMine.body);
-      }
-      return;
-    }
-    if (e.key !== 'Enter') return;
-    const shouldSend = enterToSend ? !e.shiftKey : e.shiftKey;
-    if (shouldSend) {
-      e.preventDefault();
-      if (!sending) void send();
-    }
-    // otherwise let the textarea insert a newline (default behaviour)
-  };
-
-  const load = useCallback(async () => {
-    const res = await fetch(`/api/messages?relationId=${relationId}`);
-    if (res.status === 403) { setForbidden(true); setLoading(false); return; }
-    const d = await res.json();
-    setMessages(d.messages ?? []);
-    setMentor(d.mentor ?? null);
-    setMentee(d.mentee ?? null);
-    setLoading(false);
-  }, [relationId]);
-
-  useEffect(() => { load(); }, [load]);
-  useEffect(() => { endRef.current?.scrollIntoView({ behavior: 'smooth' }); }, [messages]);
-
-  const addFiles = (list: FileList | File[]) => {
-    const picked = Array.from(list).filter(Boolean);
-    if (picked.length === 0) return;
-    setAttachments((prev) => [...prev, ...picked.map((file) => ({ file, url: URL.createObjectURL(file) }))]);
-  };
-
-  const removeAttachment = (idx: number) => {
-    setAttachments((prev) => {
-      const next = [...prev];
-      const [gone] = next.splice(idx, 1);
-      if (gone) URL.revokeObjectURL(gone.url);
-      return next;
-    });
-  };
-
-  // Paste an image straight from the clipboard into the reply box.
-  const onPaste = (e: React.ClipboardEvent) => {
-    const imgs = Array.from(e.clipboardData.items)
-      .filter((it) => it.kind === 'file' && it.type.startsWith('image/'))
-      .map((it) => it.getAsFile())
-      .filter((f): f is File => !!f)
-      // Clipboard images are often named "image.png"; make them unique.
-      .map((f) => new File([f], `pasted-${messages.length}-${f.name || 'image.png'}`, { type: f.type }));
-    if (imgs.length) { e.preventDefault(); addFiles(imgs); }
-  };
-
-  const send = async (e?: React.FormEvent) => {
-    e?.preventDefault();
-    if (!body.trim() && attachments.length === 0) return;
-    setSending(true);
-    setAttachError('');
-    try {
-      let res: Response;
-      if (attachments.length > 0) {
-        const fd = new FormData();
-        fd.append('relationId', relationId);
-        fd.append('body', body);
-        attachments.forEach((a) => fd.append('file', a.file));
-        res = await fetch('/api/messages', { method: 'POST', body: fd });
-      } else {
-        res = await fetch('/api/messages', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ relationId, body }),
-        });
-      }
-      if (res.ok) {
-        setBody('');
-        attachments.forEach((a) => URL.revokeObjectURL(a.url));
-        setAttachments([]);
-        if (fileRef.current) fileRef.current.value = '';
-        await load();
-      } else {
-        const data = await res.json().catch(() => ({}));
-        setAttachError(data.error || t.messages.sendFailed);
-      }
-    } finally {
-      setSending(false);
-    }
-  };
-
-  const react = async (id: string, emoji: string) => {
-    setReactId(null);
-    try {
-      const res = await fetch(`/api/messages/${id}/reactions`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ emoji }),
-      });
-      if (res.ok) await load();
-    } catch { /* non-critical; ignore */ }
-  };
-
-  const startEdit = (m: Msg) => { setMenuId(null); setEditId(m.id); setEditBody(m.body); };
-  const cancelEdit = () => { setEditId(null); setEditBody(''); };
-
-  const saveEdit = async (id: string) => {
-    if (!editBody.trim()) return;
-    setRowBusy(true);
-    try {
-      const res = await fetch(`/api/messages/${id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ body: editBody }),
-      });
-      if (res.ok) { cancelEdit(); await load(); }
-      else setAttachError(t.messages.editFailed);
-    } catch { setAttachError(t.messages.editFailed); }
-    finally { setRowBusy(false); }
-  };
-
-  const remove = async (id: string, scope: 'everyone' | 'me') => {
-    setMenuId(null);
-    if (scope === 'everyone' && !window.confirm(t.messages.deleteEveryoneConfirm)) return;
-    setRowBusy(true);
-    try {
-      const res = await fetch(`/api/messages/${id}?scope=${scope}`, { method: 'DELETE' });
-      if (res.ok) await load();
-      else setAttachError(t.messages.deleteFailed);
-    } catch { setAttachError(t.messages.deleteFailed); }
-    finally { setRowBusy(false); }
-  };
-
-  if (loading) return <p className="text-center py-12 text-gray-400">{t.common.loading}</p>;
-  if (forbidden) return <p className="text-center py-12 text-gray-400">{t.common.notFound}</p>;
-
-  const nameFor = (id: string) =>
-    id === mentor?.id ? mentor?.fullName : id === mentee?.id ? mentee?.fullName : '—';
-  const other = myId === mentor?.id ? mentee : mentor;
-
-  return (
-    <div>
-      <h1 className="text-2xl font-bold text-gray-900 mb-1">{t.messages.title}</h1>
-      <p className="text-gray-500 mb-6">{other?.fullName}</p>
-
-      <Card className="mb-4">
-        {messages.length === 0 ? (
-          <p className="text-center py-10 text-gray-400 text-sm">{t.messages.empty}</p>
-        ) : (
-          <div className="space-y-3 max-h-[55vh] overflow-y-auto">
-            {messages.map((m) => {
-              const mine = m.senderId === myId;
-              return (
-                <div key={m.id} className={`flex ${mine ? 'justify-end' : 'justify-start'}`}>
-                  <div className={`max-w-[80%] rounded-2xl px-4 py-2 text-sm ${mine ? 'bg-blue-600 text-white' : 'bg-gray-100 text-gray-900'}`}>
-                    {!mine && <p className="text-xs font-medium mb-0.5 opacity-70">{nameFor(m.senderId)}</p>}
-                    {m.deleted ? (
-                      <p className={`italic ${mine ? 'text-blue-100' : 'text-gray-400'}`}>{t.messages.deletedPlaceholder}</p>
-                    ) : editId === m.id ? (
-                      <div className="space-y-1.5">
-                        <textarea
-                          value={editBody}
-                          onChange={(e) => setEditBody(e.target.value)}
-                          rows={2}
-                          className="w-full rounded-lg border border-gray-300 px-2 py-1 text-sm text-gray-900 focus:outline-none focus:ring-2 focus:ring-blue-100"
-                        />
-                        <div className="flex gap-1.5">
-                          <Button size="sm" onClick={() => saveEdit(m.id)} loading={rowBusy} disabled={!editBody.trim()}>{t.messages.save}</Button>
-                          <Button size="sm" variant="outline" onClick={cancelEdit}>{t.messages.cancel}</Button>
-                        </div>
-                      </div>
-                    ) : (
-                      <>
-                        {m.body && <p className="whitespace-pre-wrap break-words">{m.body}</p>}
-                        {m.attachments.map((a) =>
-                          a.contentType.startsWith('image/') ? (
-                            <a key={a.id} href={`/api/messages/attachments/${a.id}`} target="_blank" rel="noopener noreferrer" className="block mt-1.5">
-                              <img src={`/api/messages/attachments/${a.id}`} alt={a.filename} className="max-h-48 rounded-lg" />
-                            </a>
-                          ) : (
-                            <a
-                              key={a.id}
-                              href={`/api/messages/attachments/${a.id}`}
-                              target="_blank"
-                              rel="noopener noreferrer"
-                              className={`flex items-center gap-1.5 mt-1.5 text-xs rounded-lg px-2 py-1.5 ${mine ? 'bg-blue-700 hover:bg-blue-800' : 'bg-white hover:bg-gray-50 border border-gray-200'}`}
-                            >
-                              <FileText className="h-3.5 w-3.5 flex-shrink-0" />
-                              <span className="truncate">{a.filename}</span>
-                              <Download className="h-3 w-3 flex-shrink-0 ml-auto" />
-                            </a>
-                          )
-                        )}
-                      </>
-                    )}
-                    <p className={`text-[10px] mt-1 flex items-center gap-1 ${mine ? 'text-blue-100' : 'text-gray-400'}`}>
-                      {m.channel === 'EMAIL' ? '✉ ' : ''}{formatDateTime(m.createdAt, locale)}
-                      {m.editedAt && !m.deleted && <span>· {t.messages.edited}</span>}
-                      {mine && !m.deleted && (
-                        <span
-                          className="inline-flex"
-                          title={m.readAt ? t.messages.read : t.messages.sent}
-                          aria-label={m.readAt ? t.messages.read : t.messages.sent}
-                        >
-                          {m.readAt
-                            ? <CheckCheck className="h-3.5 w-3.5 text-sky-300" aria-hidden />
-                            : <Check className="h-3.5 w-3.5" aria-hidden />}
-                        </span>
-                      )}
-                      {!m.deleted && editId !== m.id && (
-                        <span className="ml-auto flex items-center gap-0.5">
-                          <button
-                            type="button"
-                            onClick={() => { setReactId(reactId === m.id ? null : m.id); setMenuId(null); }}
-                            aria-label={t.messages.react}
-                            className={`rounded p-0.5 ${mine ? 'hover:bg-blue-700' : 'hover:bg-gray-200'}`}
-                          >
-                            <SmilePlus className="h-3.5 w-3.5" />
-                          </button>
-                          <button
-                            type="button"
-                            onClick={() => { setMenuId(menuId === m.id ? null : m.id); setReactId(null); }}
-                            aria-label={t.messages.messageActions}
-                            className={`rounded p-0.5 ${mine ? 'hover:bg-blue-700' : 'hover:bg-gray-200'}`}
-                          >
-                            <MoreVertical className="h-3.5 w-3.5" />
-                          </button>
-                        </span>
-                      )}
-                    </p>
-
-                    {/* Emoji picker */}
-                    {reactId === m.id && !m.deleted && (() => {
-                      const myEmoji = m.reactions.find((r) => r.mine)?.emoji;
-                      return (
-                        <div className="mt-1.5 flex gap-1 rounded-lg bg-white/90 border border-black/5 px-1.5 py-1 w-fit">
-                          {REACTIONS.map((emoji) => (
-                            <button
-                              key={emoji}
-                              type="button"
-                              onClick={() => react(m.id, emoji)}
-                              className={`text-base leading-none hover:scale-125 transition-transform rounded ${myEmoji === emoji ? 'ring-2 ring-blue-400 ring-offset-1 dark:ring-offset-gray-800' : ''}`}
-                              aria-label={emoji}
-                              aria-pressed={myEmoji === emoji}
-                            >
-                              {emoji}
-                            </button>
-                          ))}
-                        </div>
-                      );
-                    })()}
-
-                    {/* Reaction chips */}
-                    {!m.deleted && m.reactions.length > 0 && (
-                      <div className="mt-1.5 flex flex-wrap gap-1">
-                        {m.reactions.map((r) => (
-                          <button
-                            key={r.emoji}
-                            type="button"
-                            onClick={() => {
-                              if (r.mine) {
-                                // Open the picker so the user can switch to a different emoji
-                                // instead of immediately removing their reaction.
-                                setReactId(reactId === m.id ? null : m.id);
-                                setMenuId(null);
-                              } else {
-                                react(m.id, r.emoji);
-                              }
-                            }}
-                            aria-pressed={r.mine}
-                            className={`inline-flex items-center gap-0.5 rounded-full px-1.5 py-0.5 text-xs border ${r.mine ? 'bg-blue-100 border-blue-300 text-blue-800' : 'bg-white/90 border-black/5 text-gray-700'}`}
-                          >
-                            <span>{r.emoji}</span>
-                            <span>{r.count}</span>
-                          </button>
-                        ))}
-                      </div>
-                    )}
-                    {menuId === m.id && !m.deleted && editId !== m.id && (
-                      <div className="mt-1.5 flex flex-wrap gap-1.5">
-                        {mine && (
-                          <button type="button" onClick={() => startEdit(m)} className="text-xs px-2 py-0.5 rounded bg-white/90 text-gray-700 hover:bg-white border border-black/5">
-                            {t.messages.edit}
-                          </button>
-                        )}
-                        <button type="button" onClick={() => remove(m.id, 'me')} className="text-xs px-2 py-0.5 rounded bg-white/90 text-gray-700 hover:bg-white border border-black/5">
-                          {t.messages.deleteForMe}
-                        </button>
-                        {mine && (
-                          <button type="button" onClick={() => remove(m.id, 'everyone')} className="text-xs px-2 py-0.5 rounded bg-white/90 text-red-600 hover:bg-white border border-black/5">
-                            {t.messages.deleteForEveryone}
-                          </button>
-                        )}
-                      </div>
-                    )}
-                  </div>
-                </div>
-              );
-            })}
-            <div ref={endRef} />
-          </div>
-        )}
-      </Card>
-
-      {attachError && <p className="text-xs text-red-600 mb-2">{attachError}</p>}
-      <PendingAttachmentList attachments={attachments} onRemove={removeAttachment} removeLabel={t.common.delete} />
-      <MessageComposer
-        body={body}
-        onBodyChange={setBody}
-        onSubmit={() => void send()}
-        sending={sending}
-        hasAttachments={attachments.length > 0}
-        placeholder={t.messages.replyPlaceholder}
-        sendLabel={t.messages.send}
-        attachLabel={t.messages.attach}
-        fileInputRef={fileRef}
-        accept=".pdf,.doc,.docx,.png,.jpg,.jpeg,application/pdf,application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document,image/png,image/jpeg"
-        onFilesSelected={addFiles}
-        onPaste={onPaste}
-        onKeyDown={onComposerKeyDown}
-        inputTestId="message-attachment-input"
-      />
-      <div className="mt-1.5 flex items-center justify-between gap-3">
-        <span className="text-xs text-gray-400 truncate">{t.messages.pasteHint}</span>
-        <button
-          type="button"
-          role="switch"
-          aria-checked={enterToSend}
-          onClick={toggleEnterToSend}
-          title={enterToSend ? t.messages.enterToSendOnHint : t.messages.enterToSendOffHint}
-          className="inline-flex items-center gap-1.5 text-xs text-gray-500 hover:text-gray-700 shrink-0"
-        >
-          <span className={`relative inline-block h-4 w-7 shrink-0 rounded-full transition-colors ${enterToSend ? 'bg-blue-600' : 'bg-gray-300'}`}>
-            <span className={`absolute top-0.5 left-0.5 h-3 w-3 rounded-full bg-white transition-transform ${enterToSend ? 'translate-x-3' : 'translate-x-0'}`} />
-          </span>
-          {t.messages.enterToSend}
-        </button>
-      </div>
-    </div>
-  );
+  return <MessageThreadView target={{ kind: 'relation', id: relationId }} />;
 }
