+++ src/i18n/dictionaries.ts
@@ -1020,6 +1020,13 @@ const en = {
     enterToSendOnHint: 'Enter sends · Shift+Enter adds a new line',
     enterToSendOffHint: 'Enter adds a new line · Shift+Enter sends',
     pasteHint: 'Tip: you can paste an image, and ↑ edits your last message',
+    newChat: 'New chat',
+    newChatHint: 'People you share a project with',
+    newChatSearch: 'Search people…',
+    newChatNoMatch: 'Nobody matches that search',
+    startFailed: 'Could not start the conversation. Please try again.',
+    startNotAllowed: 'You can no longer message this person.',
+    readOnlyNotice: 'You can read this conversation but no longer send messages — you no longer share a project.',
   },
   logInteraction: {
     add: 'Log interaction',
@@ -2560,6 +2567,13 @@ const tr: Dict = {
     enterToSendOnHint: 'Enter gönderir · Shift+Enter alt satıra geçer',
     enterToSendOffHint: 'Enter alt satıra geçer · Shift+Enter gönderir',
     pasteHint: 'İpucu: resim yapıştırabilirsin, ↑ ile son mesajını düzenlersin',
+    newChat: 'Yeni sohbet',
+    newChatHint: 'Aynı projede olduğun kişiler',
+    newChatSearch: 'Kişi ara…',
+    newChatNoMatch: 'Bu aramaya uyan kimse yok',
+    startFailed: 'Sohbet başlatılamadı. Lütfen tekrar dene.',
+    startNotAllowed: 'Bu kişiye artık mesaj gönderemezsin.',
+    readOnlyNotice: 'Bu sohbeti okuyabilirsin ama artık mesaj gönderemezsin — aynı projede değilsiniz.',
   },
   logInteraction: {
     add: 'Etkileşim ekle',
@@ -4098,6 +4112,13 @@ const de: Dict = {
     enterToSendOnHint: 'Enter sendet · Umschalt+Enter fügt eine neue Zeile ein',
     enterToSendOffHint: 'Enter fügt eine neue Zeile ein · Umschalt+Enter sendet',
     pasteHint: 'Tipp: Du kannst ein Bild einfügen, ↑ bearbeitet deine letzte Nachricht',
+    newChat: 'Neuer Chat',
+    newChatHint: 'Personen, mit denen du ein Projekt teilst',
+    newChatSearch: 'Personen suchen…',
+    newChatNoMatch: 'Niemand passt zu dieser Suche',
+    startFailed: 'Die Unterhaltung konnte nicht gestartet werden. Bitte versuche es erneut.',
+    startNotAllowed: 'Du kannst dieser Person keine Nachrichten mehr senden.',
+    readOnlyNotice: 'Du kannst diese Unterhaltung lesen, aber keine Nachrichten mehr senden — ihr teilt kein Projekt mehr.',
   },
   logInteraction: {
     add: 'Interaktion erfassen',
